+++ src/alignment/mod.rs
@@ -1,17 +1,21 @@
-//! Horizontal text alignment opitons.
+//! Text alignment options.
+use crate::{rendering::cursor::Cursor, style::StyledTextBox};
+use embedded_graphics::prelude::*;
 
+pub mod bottom;
 pub mod center;
 pub mod justified;
 pub mod left;
 pub mod right;
+pub mod top;
 
-/// Text alignment base trait.
+/// Horizontal text alignment base trait.
 ///
 /// Use implementors to parametrize [`TextBoxStyle`] and [`TextBoxStyleBuilder`].
 ///
 /// [`TextBoxStyle`]: ../style/struct.TextBoxStyle.html
 /// [`TextBoxStyleBuilder`]: ../style/builder/struct.TextBoxStyleBuilder.html
-pub trait TextAlignment: Copy {
+pub trait HorizontalTextAlignment: Copy {
     /// Whether or not render spaces in the start of the line.
     const STARTING_SPACES: bool;
 
@@ -19,7 +23,26 @@ pub trait TextAlignment: Copy {
     const ENDING_SPACES: bool;
 }
 
+/// Vertical text alignment base trait.
+///
+/// Use implementors to parametrize [`TextBoxStyle`] and [`TextBoxStyleBuilder`].
+///
+/// [`TextBoxStyle`]: ../style/struct.TextBoxStyle.html
+/// [`TextBoxStyleBuilder`]: ../style/builder/struct.TextBoxStyleBuilder.html
+pub trait VerticalTextAlignment: Copy {
+    /// Set the cursor's initial vertical position
+    fn apply_vertical_alignment<'a, C, F, A>(
+        cursor: &mut Cursor<F>,
+        styled_text_box: &'a StyledTextBox<'a, C, F, A, Self>,
+    ) where
+        C: PixelColor,
+        F: Font + Copy,
+        A: HorizontalTextAlignment;
+}
+
+pub use bottom::BottomAligned;
 pub use center::CenterAligned;
 pub use justified::Justified;
 pub use left::LeftAligned;
 pub use right::RightAligned;
+pub use top::TopAligned;